Refactor project structure and move model classes
This commit moves model classes from the `Web_health_app.ApiService.Models` namespace to a new `Web_health_app.Models.Models` namespace. It adds `LoginModel` and `LoginResponseModel` classes in the new namespace and updates the `AuthController` to reference them. Project files and the solution file are modified to include a reference to the new `Models` project, which also has its own project file defining the target framework and properties.